[DX] Move UnionTypeFactory to NodeTypeResolver packages
