+++ rules/CodeQuality/NodeManipulator/ClassMethodReturnTypeManipulator.php
@@ -13,8 +13,8 @@ use PHPStan\Type\ObjectType;
 use PHPStan\Type\Type;
 use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfoFactory;
 use Rector\BetterPhpDocParser\PhpDocManipulator\PhpDocTypeChanger;
-use Rector\DowngradePhp72\UnionTypeFactory;
 use Rector\NodeTypeResolver\NodeTypeResolver;
+use Rector\NodeTypeResolver\PHPStan\Type\UnionTypeFactory;
 
 final class ClassMethodReturnTypeManipulator
 {
